refactor: move `convertMatchToParty` method to `parties.js`
