+++ lib/parties.js
@@ -93,6 +93,38 @@ export function ensurePartyIsNotRegistered(parties, match) {
 }
 
 /**
+ * Convert match from previous round into participant in the current round
+ *
+ * @param {import('./types').Match} match Match from previous round
+ * @param {number} index All registered matches in previous round
+ * @param {import('./types').Match[]} matches All registered matches in previous round
+ * @returns {import('./types').Party} Party obj based on match
+ */
+export function convertMatchToParty(match, index, matches) {
+  /** @type {import('./types').Party} */
+  const party = {
+    round: match.next.round,
+    side: match.next.side,
+    name: `Winner match ${match.id}`,
+    order: +((index + 1) / (matches.length + match.round)).toPrecision(3),
+    match: {
+      id: match.id,
+      round: match.round,
+      size: match.size,
+      span: match.next.span,
+      index,
+    },
+  }
+
+  if (match.round === 0 && match.singular) {
+    party.id = match.parties[0].id
+    party.name = match.parties[0].name
+  }
+
+  return party
+}
+
+/**
  * Determine match side for each participant
  *
  * @param {import('./types').Party[]} parties List of available participants
